raspberry/pyproject.toml
[project]
name = "raspberry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-iot-device>=2.14.0",
    "azure-iot-hub>=2.7.0",
    "python-dotenv>=1.2.1",
    "six>=1.16.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
python_files = ["test_*.py", "*test.py"]
pythonpath = ["."]
